refactor: real world object approach
