+++ src/modules/preferences/index.ts
@@ -1,5 +1,3 @@
-export * from './preferences-storage'
-export * from './use-cases/set-preference-use-case'
-export * from './use-cases/get-preference-use-case'
-export * from './models'
+export * from './storage/preferences-storage'
+export * from './preferences'
 export * from './errors'
