--- src/modules/preferences/models/index.ts
@@ -1,2 +1,0 @@
-export type Preference = 'lang'
-export const possiblePreferences: Preference[] = ['lang']
